+++ src/components/organizations/show/style.css
@@ -47,88 +47,3 @@
     }
   }
 }
-
-.OrganizationShow .MembersAddSubTab .ui.form{
-  font-size: 1.2em;
-  margin-top: 4em;
-  margin-bottom: 4em;
-}
-
-.OrganizationShow .MembersAddSubTab p{
-  color: #666;
-  font-size: 1.2em;
-  line-height: 1.7em;
-}
-
-.OrganizationShow .MembersTab {
-  margin-top: 7em;
-
-  .ui.button.green{
-    background-color: $green-2;
-  }
-
-  .members {
-    width: 100%;
-    background-color: white;
-    border-radius: 3px;
-    padding: 0 1em;
-  }
-
-  .Member {
-    padding: 1em 0.2em;
-    border-bottom: 1px solid #eee;
-
-    .role{
-      margin-top: 0.6em;
-      color: $grey-2;
-      font-weight: bold;
-      font-size: 1.2em;
-    }
-
-    .button.remove {
-      margin-top: 0.2em;
-      float: right;
-      font-size: 1.2em !important;
-      padding: 7px 10px !important;
-    }
-
-    .button.remove i{
-      margin: 0 !important;
-    }
-
-    .name {
-      float: left;
-      font-size: 1.3em;
-      font-weight: 800;
-      color: $black-3;
-      margin-top: 0.5em;
-    }
-
-    .invitation-status {
-      font-size: 1.2em;
-      color: #666;
-      margin-top: 0.6em;
-    }
-
-    .avatar {
-      float: left;
-      width: 40px;
-      border-radius: 3px;
-      margin-right: 1em;
-      font-size: 2.9em;
-      margin-left: 0.3em;
-      color: $blue-5;
-      text-align: center;
-    }
-
-    .avatar .fa{
-      font-size: 0.6em;
-      margin-top: 5px;
-      opacity: 0.7;
-    }
-
-    &:last-child {
-      border-bottom: none;
-    }
-  }
-}
