+++ src/modules/newOrg/reducer.js
@@ -1,0 +1,10 @@
+export function newOrgR(state = {}, action) {
+  switch (action.type) {
+    case 'CLEAR_NEW_ORGANIZATION':
+      return {}
+    case 'ORGANIZATIONS_CREATE_SUCCESS':
+      return action.record
+    default:
+      return state
+  }
+}
